[Refactor]: Core의 ServerSchemas 통합돼있던 내용 Enums, Request, Response 로 분리
