--- Assets/Scripts/Core/ServerSchemas.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 5e4ffcde463a14674881c8bc789db9d0
