+++ Assets/Scripts/Network/DTOs/User.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8173a8c702d3e45cab6d33ed4a6635ef
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
