+++ Assets/Scripts/Network/DTOs/User/UserEnums.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 6583b5a47dc844ab893e1952b9202ce5
