+++ Assets/Scripts/Network/DTOs/User/UserResponse.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 93b7d44fdc5c246c08cc5da1cdd162f8
